refactor: extract shared useDialogFocusTrap hook for overlays (#206)
