+++ src/components/features/execute/OwrExportOverlay.tsx
@@ -1,7 +1,8 @@
-import { useEffect, useId, useRef, useState } from "react";
+import { useId, useState } from "react";
 import { Button } from "@/components/ui/Button";
 import { Card, CardContent, CardHeader } from "@/components/ui/Card";
 import { Label } from "@/components/ui/Label";
+import { useDialogFocusTrap } from "@/lib/hooks/useDialogFocusTrap";
 import {
   X,
   Globe,
@@ -63,71 +64,7 @@ export function OwrExportOverlay({
 }: OwrExportOverlayProps) {
   const [isOwrCopied, setIsOwrCopied] = useState(false);
   const titleId = useId();
-  const dialogRef = useRef<HTMLDivElement>(null);
-  const closeButtonRef = useRef<HTMLButtonElement>(null);
-
-  // Focus management: move focus in on open, trap Tab, Escape to close, restore on close.
-  useEffect(() => {
-    if (!open) return;
-
-    const previouslyFocused =
-      document.activeElement instanceof HTMLElement ? document.activeElement : null;
-
-    const focusInitial = () => {
-      closeButtonRef.current?.focus();
-      if (document.activeElement !== closeButtonRef.current) {
-        dialogRef.current?.focus();
-      }
-    };
-    const focusTimer = window.setTimeout(focusInitial, 0);
-
-    const getFocusable = (): HTMLElement[] => {
-      const root = dialogRef.current;
-      if (!root) return [];
-      return Array.from(
-        root.querySelectorAll<HTMLElement>(
-          'button:not([disabled]), [href], input:not([disabled]), select:not([disabled]), textarea:not([disabled]), [tabindex]:not([tabindex="-1"])',
-        ),
-      ).filter((el) => !el.hasAttribute("disabled") && el.offsetParent !== null);
-    };
-
-    const onKeyDown = (event: KeyboardEvent) => {
-      if (event.key === "Escape") {
-        event.preventDefault();
-        onClose();
-        return;
-      }
-      if (event.key !== "Tab") return;
-
-      const focusable = getFocusable();
-      if (focusable.length === 0) {
-        event.preventDefault();
-        dialogRef.current?.focus();
-        return;
-      }
-
-      const first = focusable[0]!;
-      const last = focusable[focusable.length - 1]!;
-      const active = document.activeElement;
-
-      if (event.shiftKey) {
-        if (active === first || !dialogRef.current?.contains(active)) {
-          event.preventDefault();
-          last.focus();
-        }
-      } else if (active === last) {
-        event.preventDefault();
-        first.focus();
-      }
-    };
-
-    window.addEventListener("keydown", onKeyDown);
-    return () => {
-      window.clearTimeout(focusTimer);
-      window.removeEventListener("keydown", onKeyDown);
-      previouslyFocused?.focus();
-    };
-  }, [open, onClose]);
+  const { dialogRef, closeButtonRef } = useDialogFocusTrap(open, onClose);
 
   if (!open) return null;
